Create Search Insert Position
Given a sorted array of distinct integers and a target value, return the index if the target is found. If not, return the index where it would be if it were inserted in order.
Output Status:
Runtime: 4 ms, faster than 80.10% of C++ online submissions for Search Insert Position.
Memory Usage: 9.7 MB, less than 46.24% of C++ online submissions for Search Insert Position.