Create Equilibrium Point.cpp
Given an array A of n positive numbers. The task is to find the first Equilibium Point in the array. 
Equilibrium Point in an array is a position such that the sum of elements before it is equal to the sum of elements after it.
Output Status:
Correct Answer.Correct Answer
Execution Time:0.31