Create Armstrong Numbers.cpp
or a given 3 digit number, find whether it is armstrong number or not. An Armstrong number of three digits is an integer such that the sum of the cubes of its digits is equal to the number itself. Return "Yes" if it is a armstrong number else return "No".
NOTE: 371 is an Armstrong number since 3^3 + 7^3 + 1^3 = 371
Output Status:
Correct Answer.Correct Answer
Execution Time:0.01